+++ src/nmCmdList.c
@@ -52,7 +52,7 @@ extern u64 FUN_003a4360_typed(u64 param_1,long param_2);
 #pragma alias FUN_003a27c0_typed FUN_003a27c0
 extern void FUN_003a27c0_typed(int param_1,s16 param_2);
 #pragma alias FUN_003a2830_nm_typed FUN_003a2830
-extern s16 FUN_003a2830_nm_typed(int param_1);
+extern u32 FUN_003a2830_nm_typed(int param_1);
 extern u8 DAT_0095be90_abs[];
 #pragma alias FUN_003c7d50_ptr FUN_003c7d50
 extern u8 *FUN_003c7d50_ptr(void);
@@ -1804,7 +1804,7 @@ void FUN_003c7560(u8 param_1)
 }
 #define FUN_003c7560(...) ((void (*)(...))FUN_003c7560)(__VA_ARGS__)
 #undef FUN_003c75b0
-// FUN_003C75B0 NONMATCHING
+// FUN_003C75B0
 
 
 u32 FUN_003c75b0(void)
